requirements.txt
requests>=2.28.0
PyJWT>=2.6.0
Pillow>=9.4.0